cleanup the course page with components
