+++ src/lib/ratings.ts
@@ -28,7 +28,7 @@ const CourseRatingsQuery = `
   }
 `
 
-export async function loadRatings(slug: string, type: string) {
+export async function loadRatings(slug: string, type: string = 'Series') {
   const SIZE_OF_PAGE = 3
 
   console.log({slug, type})
@@ -46,6 +46,10 @@ export async function loadRatings(slug: string, type: string) {
     ...authorizationHeader,
   })
 
-  const result = await graphQLClient.request(CourseRatingsQuery, variables)
-  return get(result, 'ratings.data', [])
+  try {
+    const result = await graphQLClient.request(CourseRatingsQuery, variables)
+    return get(result, 'ratings.data', [])
+  } catch {
+    return []
+  }
 }
